Refactor(global): member 세션을 얻을 때 const상수 사용
LOGIN_MEMBER 상수를 사용해 세션 "member"부분을 가져올 수 있습니다.
