+++ src/main/java/com/example/newsfeed/comment/service/CommentService.java
@@ -4,7 +4,7 @@ import com.example.newsfeed.comment.dto.CommentRequestDto;
 import com.example.newsfeed.comment.dto.CommentResponseDto;
 import com.example.newsfeed.comment.entity.Comment;
 import com.example.newsfeed.comment.repository.CommentRepository;
-import com.example.newsfeed.global.entity.SessionMemberDto;
+import com.example.newsfeed.global.dto.SessionMemberDto;
 import com.example.newsfeed.member.entity.Member;
 import com.example.newsfeed.post.entity.Post;
 import com.example.newsfeed.post.service.PostService;
