+++ src/main/java/com/example/newsfeed/follow/controller/FollowController.java
@@ -2,7 +2,7 @@ package com.example.newsfeed.follow.controller;
 
 import com.example.newsfeed.follow.service.FollowService;
 import com.example.newsfeed.global.annotation.LoginRequired;
-import com.example.newsfeed.global.entity.SessionMemberDto;
+import com.example.newsfeed.global.dto.SessionMemberDto;
 import com.example.newsfeed.member.dto.response.MemberListGetResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,6 +11,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static com.example.newsfeed.global.constant.SessionConst.LOGIN_MEMBER;
 
 @Slf4j
 @RestController
@@ -24,7 +26,7 @@ public class FollowController {
     @PostMapping("/{memberId}")
     public ResponseEntity<Void> createFollow(
             @PathVariable Long memberId,
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         followService.createFollow(session.getId(), memberId);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -34,7 +36,7 @@ public class FollowController {
     @DeleteMapping("/{memberId}")
     public ResponseEntity<Void> deleteFollow(
             @PathVariable Long memberId,
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         followService.deleteFollow(session.getId(), memberId);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -43,7 +45,7 @@ public class FollowController {
     @LoginRequired
     @GetMapping
     public ResponseEntity<List<MemberListGetResponseDto>> findAllFollowerMembers(
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         List<MemberListGetResponseDto> responseDtoList = followService.findAllFollowerMembers(session.getId());
         return new ResponseEntity<>(responseDtoList, HttpStatus.OK);
