+++ src/main/java/com/example/newsfeed/global/interceptor/LoginInterceptor.java
@@ -1,7 +1,7 @@
 package com.example.newsfeed.global.interceptor;
 
 import com.example.newsfeed.global.annotation.LoginRequired;
-import com.example.newsfeed.global.entity.SessionMemberDto;
+import com.example.newsfeed.global.dto.SessionMemberDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
