+++ src/main/java/com/example/newsfeed/like/controller/CommentLikeController.java
@@ -1,12 +1,14 @@
 package com.example.newsfeed.like.controller;
 
 import com.example.newsfeed.global.annotation.LoginRequired;
-import com.example.newsfeed.global.entity.SessionMemberDto;
+import com.example.newsfeed.global.dto.SessionMemberDto;
 import com.example.newsfeed.like.service.CommentLikeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import static com.example.newsfeed.global.constant.SessionConst.LOGIN_MEMBER;
 
 @RestController
 @RequestMapping("/comments")
@@ -19,7 +21,7 @@ public class CommentLikeController{
     @PostMapping("/{commentId}/likes")
     public ResponseEntity<Void> createLike(
             @PathVariable Long commentId,
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         commentLikeService.createLike(session.getId(), commentId);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -29,7 +31,7 @@ public class CommentLikeController{
     @DeleteMapping("/{commentId}/likes")
     public ResponseEntity<Void> deleteLike(
             @PathVariable Long commentId,
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         commentLikeService.deleteLike(session.getId(), commentId);
         return new ResponseEntity<>(HttpStatus.OK);
