+++ src/main/java/com/example/newsfeed/like/controller/PostLikeController.java
@@ -1,12 +1,14 @@
 package com.example.newsfeed.like.controller;
 
 import com.example.newsfeed.global.annotation.LoginRequired;
-import com.example.newsfeed.global.entity.SessionMemberDto;
+import com.example.newsfeed.global.dto.SessionMemberDto;
 import com.example.newsfeed.like.service.PostLikeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import static com.example.newsfeed.global.constant.SessionConst.LOGIN_MEMBER;
 
 @RestController
 @RequestMapping("/posts")
@@ -19,7 +21,7 @@ public class PostLikeController{
     @PostMapping("/{postId}/likes")
     public ResponseEntity<Void> createLike(
             @PathVariable Long postId,
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         postLikeService.createLike(session.getId(), postId);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -29,7 +31,7 @@ public class PostLikeController{
     @DeleteMapping("/{postId}/likes")
     public ResponseEntity<Void> deleteLike(
             @PathVariable Long postId,
-            @SessionAttribute(name = "member") SessionMemberDto session
+            @SessionAttribute(name = LOGIN_MEMBER) SessionMemberDto session
     ) {
         postLikeService.deleteLike(session.getId(), postId);
         return new ResponseEntity<>(HttpStatus.OK);
