+++ src/main/java/com/example/newsfeed/post/dto/PostResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.newsfeed.post.dto;
 
-import com.example.newsfeed.global.entity.SessionMemberDto;
+import com.example.newsfeed.global.dto.SessionMemberDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import java.time.LocalDateTime;
